+++ src/main/gen/com/github/xepozz/robots_txt/language/psi/RobotsTxtDirective.java
@@ -1,0 +1,8 @@
+// This is a generated file. Not intended for manual editing.
+package com.github.xepozz.robots_txt.language.psi;
+
+import com.intellij.psi.PsiElement;
+
+public interface RobotsTxtDirective extends PsiElement {
+
+}