+++ src/main/gen/com/github/xepozz/robots_txt/language/psi/RobotsTxtRule.java
@@ -1,10 +1,16 @@
 // This is a generated file. Not intended for manual editing.
 package com.github.xepozz.robots_txt.language.psi;
 
-import java.util.List;
-import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public interface RobotsTxtRule extends PsiElement {
+
+    @NotNull
+    RobotsTxtDirective getDirective();
+
+    @Nullable
+    RobotsTxtValue getValue();
 
 }