+++ src/main/gen/com/github/xepozz/robots_txt/language/psi/RobotsTxtVisitor.java
@@ -1,11 +1,15 @@
 // This is a generated file. Not intended for manual editing.
 package com.github.xepozz.robots_txt.language.psi;
 
-import org.jetbrains.annotations.*;
-import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import org.jetbrains.annotations.NotNull;
 
 public class RobotsTxtVisitor extends PsiElementVisitor {
+
+    public void visitDirective(@NotNull RobotsTxtDirective o) {
+        visitPsiElement(o);
+    }
 
   public void visitEntry(@NotNull RobotsTxtEntry o) {
     visitPsiElement(o);
@@ -14,6 +18,10 @@ public class RobotsTxtVisitor extends PsiElementVisitor {
   public void visitRule(@NotNull RobotsTxtRule o) {
     visitPsiElement(o);
   }
+
+    public void visitValue(@NotNull RobotsTxtValue o) {
+        visitPsiElement(o);
+    }
 
   public void visitPsiElement(@NotNull PsiElement o) {
     visitElement(o);