+++ src/main/kotlin/com/github/xepozz/robots_txt/ide/RobotsTxtCommenter.kt
@@ -1,0 +1,15 @@
+package com.github.xepozz.robots_txt.ide
+
+import com.intellij.lang.Commenter
+
+class RobotsTxtCommenter : Commenter {
+    override fun getLineCommentPrefix() = "#"
+
+    override fun getBlockCommentPrefix() = null
+
+    override fun getBlockCommentSuffix() = null
+
+    override fun getCommentedBlockCommentPrefix() = null
+
+    override fun getCommentedBlockCommentSuffix() = null
+}